+++ blog-article/src/main/java/com/mrqinzh/article/mapper/ArticleMapper.java
@@ -14,9 +14,9 @@ public interface ArticleMapper extends BaseMapper<Article> {
     long pageCount(PageDTO pageDTO);
     List<Article> list(PageDTO pageDTO); // 排序全部文章、以及分页
 
-    Article getById(Integer articleId); // 根据文章id展示当前文章
+    Article getById(Long articleId); // 根据文章id展示当前文章
 
     @Update("update article set status = 1 where id = #{articleId}")
-    Boolean deleteStatus(Integer articleId); // 删除一篇文章
+    Boolean deleteStatus(Long articleId); // 删除一篇文章
 
 }
